Fix delete book

+++ src/pages/Admin/AdminBookListView/AdminBookListView.jsx
@@ -2,16 +2,18 @@ import './AdminBookListView.scss';
 import EditIcon from '@/assets/icons/EditIcon.svg';
 import DeleteIcon from '@/assets/icons/DeleteIcon.svg';
 import NodeIcon from '@/assets/icons/BookIcon.svg';
-import { useEffect, useState } from 'react';
+import { useContext, useEffect, useState } from 'react';
 import { NavLink } from 'react-router-dom';
 import AddIcon from '@/assets/icons/AddIcon.svg';
 import { BaseButton } from '@/composants/Base/BaseButton/BaseButton.jsx';
+import { AuthContext } from '@/composants/AuthContext/AuthContext.jsx';
 
 export function AdminBookListView() {
   const [, setSearch] = useState('');
   const [books, setBooks] = useState([]);
 
   const apiURL = import.meta.env.VITE_API_URL || 'http://193.168.146.103:3000';
+  const { user } = useContext(AuthContext);
 
   useEffect(() => {
     // Fetch books from the server
@@ -54,8 +56,13 @@ export function AdminBookListView() {
   }
 
   async function handleDeleteBook(id) {
+    if (!user) return;
     await fetch(`${apiURL}/livres/${id}`, {
-      method: 'DELETE'
+      method: 'DELETE',
+      headers: {
+        'Content-Type': 'application/json',
+        Authorization: user.token
+      }
     }).then((response) => {
       if (response.ok) {
         // Remove the book from the list
